Start cleaning up texlib a bit
The paragraphs it produces in some “Lorem ipsum” cases look bad enough
that I worried the algorithm was broken, so here are some tweaks and
improvements that — to me — make it more readable.  But the paragraphs
still look bad.  Not sure if it’s just a bad paragraph, or what.